Make `Result` a wrapper, not a re-implementation.
Previously, `propagate::Result` was mostly a copy of
`std::result::Result`, with a different `Err` type and modified
`FromResidual` implementations.

Because of this, users would have to to use
`propagate::Result::{Ok, Err}` to construct results (unless they used
`try blocks`).

This changes `propagate::Result<T, E, S>` to be a minimal wrapper around
a `std::result::Result<T, TracedError<E, S>>`. If users want to get to
the wrapped result, they can call `.unpack()`, which will also return
the traced stack if there is one.

With this change, users no longer have to (nor have the ability to)
shadow the standard library's `Ok` and `Err`.
Now, there should be much less confusion about what a
`propagate::Result` really is. It no longer pretends to be the same
thing as a `std::result::Result`.